+++ components/src/main/java/com/mercadolibre/android/andesui/amountfield/utils/AndesAmountFieldIntFormatter.kt
@@ -189,8 +189,8 @@ internal class AndesAmountFieldIntFormatter(
      * Call after finishing the editable work.
      */
     private fun close(editable: Editable) {
-        amountListener.onAmountChanged(isExceeded)
         resizingListener.resizeComponentIfNeeded()
+        amountListener.onAmountChanged(isExceeded)
         editable.filters = filters
         isRunning = false
     }